feat: got links working, got named files correctly named, removed popout and link. todo: fix names for url slug
