+++ package.json
@@ -24,7 +24,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.11.1",
-    "reactjs-popup": "^2.0.5",
     "sort-by": "^1.2.0"
   },
   "devDependencies": {
