--- src/mdx-components/link.mdx
@@ -1,17 +1,0 @@
-import JSXPopout from "~/components/popout"
-import { getJsxNameFromRelativePath } from "~/util/get-jsx-name";
-
-export default function Popout(props) {
-    const fileName = props.children._source.fileName
-    const trigger = getJsxNameFromRelativePath(fileName)
-
-    // @TODO - this is broken! Just an attribute for the trigger or component, like the imported popout module / an a href
-    const PopoutTrigger = () => (<a>{trigger}</a>)
-
-    return (
-        <JSXPopout trigger={<PopoutTrigger />}>
-            {props.children}
-        </JSXPopout>
-    )
-
-}
