--- src/mdx-components/popout.mdx
@@ -1,16 +1,0 @@
-import JSXPopout from "~/components/popout";
-import { getJsxNameFromRelativePath } from "~/util/get-jsx-name";
-
-export default function Popout(props) {
-    const fileName = props.children._source.fileName
-    const trigger = getJsxNameFromRelativePath(fileName)
-
-    const PopoutTrigger = () => (<a>{trigger}</a>)
-
-    return (
-        <JSXPopout trigger={<PopoutTrigger />}>
-            {props.children}
-        </JSXPopout>
-    )
-
-}
