+++ src/util/name/get-jsx-name.ts
@@ -1,0 +1,13 @@
+import camelCase from "lodash/camelCase";
+import {
+  capitalizeFirstLetter,
+  removeRelativePathFromFilename,
+} from "../string-helpers";
+
+export function getJsxNameFromRelativePath(relativepath: string): string {
+  return getJsxName(removeRelativePathFromFilename(relativepath));
+}
+
+function getJsxName(filename: string): string {
+  return capitalizeFirstLetter(camelCase(filename));
+}
