+++ src/util/name/get-url-slug-name.ts
@@ -1,0 +1,10 @@
+import { removeRelativePathFromFilename } from "../string-helpers";
+import kebabCase from "lodash/kebabCase";
+
+export function getUrlSlugNameFromRelativePath(relativepath: string): string {
+  return getUrlSlugName(removeRelativePathFromFilename(relativepath));
+}
+
+function getUrlSlugName(filename: string): string {
+  return kebabCase(filename);
+}
