+++ src/util/name/index.ts
@@ -1,0 +1,2 @@
+export * from "./get-url-slug-name"
+export * from "./get-jsx-name"
